Refactor: 추천 메뉴 컴포저블 구조 분리 및 UI 개선
`RecommendedCard`에 포함되어 있던 하위 컴포저블들을 별도의 파일로 분리하여 `recommended` 패키지 내에서 관리하도록 구조를 개선했습니다. 또한, 장바구니 화면의 전반적인 UI를 수정했습니다.

- **`RecommendedCard.kt`**
    - `RecommendedRow`와 `RecommendedAddButton` 컴포저블을 각각 `RecommendedRow.kt`, `RecommendedButton.kt` 파일로 분리했습니다.
    - 기존에 `RecommendedCard` 내에 있던 "함께 먹으면 좋아요" 텍스트를 `CartScreen`으로 이동시켜, 카드의 재사용성을 높였습니다.
    - `RecommendedCard`에 `gray300` 색상의 테두리(border)를 추가했습니다.

- **`CartScreen.kt`**
    - `RecommendedCard` 위에 "함께 먹으면 좋아요" 제목 텍스트를 추가하고, 상하 간격을 조정했습니다.
    - 파일 경로 변경에 따라 `RecommendCard`의 import 경로를 수정했습니다.

- **`MenuAddButton.kt`**
    - "메뉴 더 담기" 버튼에 수직 패딩(vertical padding)을 추가하여 UI를 개선했습니다.
